fix!: don't checkout ref every time
If there is a commit above yours and you are re-running an action, this checkout will always checkout the `main` ref which will actually be a different SHA than the one that triggered the action in the first place. If you are re-running an action, we should always use the same SHA that triggered the action. BTW, this is the default behavior of [actions/checkout](https://github.com/actions/checkout), we're just writing it out specifically so that you can overwrite it with a input parameter.